Updated code based on 1st Udacity review
I added a try+except block at the top of the db file, to handle any errors when connecting to the database.
I also remove all but one of the connections to the db to improve performance.
The one and only connection to the db is a globally defined variable in the try+except block, and the connection closes within the last function.